Edit the returned values to include P_dot, update COMPAS function name
I added P_dot to the values returned by the simulation. These values are calculated by the calc_NS_spindown_power.

I updated all the calls of the function previously named RLO_evolve_COMPAS to RLO_evolve_Chattopadhyay2020. This is the new name for the function in pulsar_modeling.py.